dividida classe janelaPesquisa em duas classes diferentes e adaptado código
